Create 334.py
Find out n2 where storing smallest value where smaller value on the left has been found. If we find value bigger than n2 on the right it is triplet.
Traverse from left to right
If we find bigger value than n2 we find triplet
We change n1 when we find smaller value than n1. Or if we find n where bigger than n1 and smaller than n2 we change n2 as n. (this will keep n2 as if we have find smaller than n1 but n2 before n2 will still persist value but if we find n2 with smaller value it will update with smaller value)