Create 1372.py
From each node, we track if it comes from left/right from the parent node and the current length of the path. We return the maximum path where the current node is included.
The maximum path will be either going from ancestor to opposite direct from the parent node or starting from current nodeto the same direction as the parent. 
By doing recursion, we will compute the max value of the length.